+++ lib/aruba/api/command.rb
@@ -122,22 +122,37 @@ module Aruba
       # @param [Integer] timeout
       #   If the timeout is reached the command will be killed
       #
+      # @param [String] stop_signal
+      #   Use signal to stop command (Private)
+      #
       # @yield [SpawnProcess]
       #   Run block with process
       #
       # rubocop:disable Metrics/MethodLength
-      def run(cmd, exit_timeout = nil, io_wait_timeout = nil)
+      # rubocop:disable Metrics/CyclomaticComplexity
+      def run(cmd, exit_timeout = nil, io_wait_timeout = nil, stop_signal = nil, startup_wait_time = nil)
         exit_timeout ||= aruba.config.exit_timeout
         io_wait_timeout ||= aruba.config.io_wait_timeout
+        stop_signal ||= aruba.config.stop_signal
+        startup_wait_time ||= aruba.config.startup_wait_time
+
+        cmd = replace_variables(cmd)
 
         @commands ||= []
         @commands << cmd
 
+        environment       = aruba.environment.to_h
+        working_directory = expand_path('.')
+
+        announcer.announce(:full_environment, environment)
+        announcer.announce(:timeout, 'exit', exit_timeout)
+        announcer.announce(:timeout, 'io wait', io_wait_timeout)
+        announcer.announce(:wait_time, 'startup wait time', startup_wait_time)
+
+        announcer.announce(:directory, working_directory)
         announcer.announce(:command, cmd)
 
         cmd               = Aruba.platform.detect_ruby(cmd)
-        environment       = aruba.environment.to_h
-        working_directory = expand_path('.')
 
         mode = if Aruba.process
                  # rubocop:disable Metrics/LineLength
@@ -157,11 +172,6 @@ module Aruba
                        aruba.config.main_class
                      end
 
-        announcer.announce(:directory, working_directory)
-        announcer.announce(:timeout, 'exit', exit_timeout)
-        announcer.announce(:timeout, 'io wait', io_wait_timeout)
-        announcer.announce(:full_environment, environment)
-
         command = Command.new(
           cmd,
           :mode              => mode,
@@ -169,7 +179,9 @@ module Aruba
           :io_wait_timeout   => io_wait_timeout,
           :working_directory => working_directory,
           :environment       => environment,
-          :main_class        => main_class
+          :main_class        => main_class,
+          :stop_signal       => stop_signal,
+          :startup_wait_time => startup_wait_time
         )
 
         if aruba.config.before? :cmd
@@ -184,10 +196,13 @@ module Aruba
         process_monitor.register_process(cmd, command)
         command.start
 
+        announcer.announce(:stop_signal, command.pid, stop_signal) if stop_signal
+
         aruba.config.after(:command, self, command)
 
         block_given? ? yield(command) : command
       end
+      # rubocop:enable Metrics/CyclomaticComplexity
       # rubocop:enable Metrics/MethodLength
 
       # Run a command with aruba