+++ lib/aruba/cucumber/command.rb
@@ -1,3 +1,7 @@
+if Aruba::VERSION < '1.0.0'
+  require 'aruba/cucumber/core'
+end
+
 When(/^I run "(.*)"$/)do |cmd|
   warn(%{\e[35m    The /^I run "(.*)"$/ step definition is deprecated. Please use the `backticks` version\e[0m})
 
@@ -35,6 +39,11 @@ When(/^I run `([^`]*)` interactively$/)do |cmd|
   @interactive = run(cmd)
 end
 
+# Merge interactive and background after refactoring with event queue
+When(/^I run `([^`]*)` in background$/)do |cmd|
+  run(sanitize_text(cmd))
+end
+
 When(/^I type "([^"]*)"$/) do |input|
   type(unescape_text(input))
 end
@@ -47,6 +56,29 @@ When(/^I pipe in (?:a|the) file(?: named)? "([^"]*)"$/) do |file|
   pipe_in_file(file)
 
   close_input
+end
+
+When(/^I (terminate|stop) the command (?:"([^"]*)"|(?:started last))$/) do |signal, command|
+  if command
+    cmd = all_commands.find { |c| c.commandline == command }
+    fail ArgumentError, %(No command "#{command}" found) if cmd.nil?
+
+    if signal == 'terminate'
+      # last_command_started.terminate
+      process_monitor.terminate_process!(process_monitor.get_process(command))
+    else
+      # last_command_started.stop
+      process_monitor.stop_process(process_monitor.get_process(command))
+    end
+  else
+    if signal == 'terminate'
+      # last_command_started.terminate
+      process_monitor.terminate_process!(last_command_started)
+    else
+      # last_command_started.stop
+      process_monitor.stop_process(last_command_started)
+    end
+  end
 end
 
 When(/^I stop the command(?: started last)? if (output|stdout|stderr) contains:$/) do |channel, expected|
@@ -330,5 +362,48 @@ Then(/^(?:the )?(output|stdout|stderr) should( not)? contain all of these lines:
     else
       expect(all_commands).to include_an_object have_output an_output_string_including(expected)
     end
+  end
+end
+
+Given(/the default aruba timeout is (\d+) seconds/) do |seconds|
+  # rubocop:disable Metrics/LineLength
+  Aruba.platform.deprecated(%{The /^the default aruba timeout is (\d+) seconds/ step definition is deprecated. Please use /^the default aruba exit timeout is (\d+) seconds/ step definition is deprecated.})
+  # rubocop:enable Metrics/LineLength
+
+  aruba.config.exit_timeout = seconds.to_i
+end
+
+Given(/The default aruba timeout is (\d+) seconds/) do |seconds|
+  # rubocop:disable Metrics/LineLength
+  Aruba.platform.deprecated(%{The /^The default aruba timeout is (\d+) seconds/ step definition is deprecated. Please use /^the default aruba exit timeout is (\d+) seconds/ step definition is deprecated.})
+  # rubocop:enable Metrics/LineLength
+
+  aruba.config.exit_timeout = seconds.to_i
+end
+
+Given(/^the (?:default )?aruba io wait timeout is (\d+) seconds?$/) do |seconds|
+  aruba.config.io_wait_timeout = seconds.to_i
+end
+
+Given(/^the (?:default )?aruba exit timeout is (\d+) seconds?$/) do |seconds|
+  aruba.config.exit_timeout = seconds.to_i
+end
+
+Given(/^the (?:default )?aruba stop signal is "([^"]*)"$/) do |signal|
+  aruba.config.stop_signal = signal
+end
+
+Given(/^I wait (\d+) seconds? for (?:a|the) command to start up$/) do |seconds|
+  aruba.config.startup_wait_time = seconds.to_i
+end
+
+When(/^I send the signal "([^"]*)" to the command (?:"([^"]*)"|(?:started last))$/) do |signal, command|
+  if command
+    cmd = all_commands.find { |c| c.commandline == command }
+    fail ArgumentError, %(No command "#{command}" found) if cmd.nil?
+
+    cmd.send_signal signal
+  else
+    last_command_started.send_signal signal
   end
 end